Extract a common subclass
